rename navigation component and create graph component
