+++ src/App.js
@@ -1,70 +1,23 @@
-import React, { Component } from 'react';
+import React from 'react';
 import './App.css';
-import { Navigation } from './components/navigation/navigation.js';
-// import { Data } from './components/data/data.js';
+import { Navigation } from './components/Navigation/Navigation.js';
+import { GraphComponent } from './components/GraphComponent/GraphComponent.js';
 
 import { useQuery } from '@apollo/react-hooks';
 import gql from "graphql-tag";
 
 import CanvasJSReact from './canvasjs.react';
 //var CanvasJSReact = require('./canvasjs.react');
-var CanvasJS = CanvasJSReact.CanvasJS;
+// var CanvasJS = CanvasJSReact.CanvasJS;
 var CanvasJSChart = CanvasJSReact.CanvasJSChart;
 
 
 function App() {
-  // get query data
-  const { data, loading, error } = useQuery(GET_RESULT);
-
-  console.log(data);
-  var confirmed_array = [];
-  var death_case_array = [];
-
-  var list = data && data.results && data.results.map((result) => {
-    confirmed_array.push({ y: result.confirmed, label: result.date })
-    death_case_array.push({ y: result.deaths, label: result.date })
-  });
-  
-  console.log(death_case_array)
-
-  // create/initialize graph
-  const options = {
-    animationEnabled: true,	
-    title:{
-      text: "Covid 19 Data"
-    },
-    height:420,
-    width:1200,
-    axisY : {
-      title: "NumberOfCases",
-      includeZero: true
-    },
-    toolTip: {
-      shared: true
-    },
-    data: [{
-      type: "spline",
-      name: "confirmed",
-      showInLegend: true,
-      dataPoints: confirmed_array
-    },
-    {
-      type: "spline",
-      name: "death",
-      showInLegend: true,
-      dataPoints: death_case_array
-    }]
-  }
-
   return (
     <div className="App">
         <Navigation />
+        <GraphComponent />
 
-        <div>
-        <CanvasJSChart options = {options}
-            /* onRef = {ref => this.chart = ref} */
-        />
-      </div>
         {/* <div className="container">
           {data &&
             data.results &&
@@ -85,19 +38,3 @@ function App() {
 }
 
 export default App;
-
-
-const GET_RESULT = gql`
-{
-  results (countries: "Canada", date: { lt: "4/9/2020" }) {
-    country {
-      name
-    }
-    date
-    confirmed
-    deaths
-    recovered
-    growthRate
-  }
-}
-`
